add home feature
